Fix offline packaging tasks, add backcompat tasks
