+++ .gitignore
@@ -19,3 +19,5 @@ coverage/
 
 \.DS_Store
 vsix/
+
+\.vscodeignore
